"1) Счетчик должен увеличиваться автоматически, его не нужно передавать в конструктор. Для этого его нужно сделать статическим, увеличивать на 1 в конструкторе и проставлять в id для каждого объекта в employeeID. 2) printAvarageSalaryEmployee возвращает просто сумму всез зарплат. 3) Лучше создать отдельный Main класс, в котором будут реализованы все методы и создаваться сотрудники." - вроде все поправил
